add logger, basic initial test case
