+++ app/__init__.py
@@ -6,7 +6,7 @@ from app.constants import CONFIG_JSON_PATH
 def create_app():
     app = FastAPI()
 
-    from app.api.translateAPI import translate
+    from app.views.v1.translate import translate
     app.include_router(translate)
 
     @app.on_event('startup')
